refactor(policy): move SelfHealing doctrine to policy-owned src/policy/ (sage #464)
Addresses sage's [important] finding: projection doctrine content belongs in the
policy layer, not src/adapters/shared/. Moves self-healing-doctrine.ts to
src/policy/; shared/index.ts re-exports it so the 6 adapters are unchanged. Done
atomically (move + import-fix together) — relocating a module soma's own
policy-guard imports is otherwise fragile.

Verify: tsc --noEmit clean; bun test 2049 pass / 0 fail; doctrine test 3 pass.
